requirements.txt
asgiref==3.8.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.3
django-bootstrap5==23.4
django-dynamic-preferences==1.16.0
django-filter==24.1
django-tables2==2.7.0
django-tinymce==3.7.1
idna==3.6
persisting-theory==1.0
pillow==10.2.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.1
